wip 関数を一つのファイルにまとめていたのを分離
