+++ app/_features/factory/components/factoryItem.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import React, { FC } from 'react'
 import { UserIcon } from '@heroicons/react/24/solid'
-import { getReviewsByFactoryId } from '../../review/api'
+import { getReviewsByFactoryId } from '../../review/api/getReviewsByFactoryId'
 
 type Props = {
   factory: Factory
