+++ app/_features/review/api/getReviewsByUserId.ts
@@ -1,0 +1,11 @@
+import { Review } from '@/app/_common/types'
+
+export const getReviewsByUserId = async (userId: string): Promise<Review[]> => {
+  try {
+    const res = await fetch(`${process.env.API_URL}/review/user/${userId}`)
+    const reviews = await res.json()
+    return reviews
+  } catch (error: any) {
+    throw new Error(error)
+  }
+}
