+++ app/api/review/getReviewByUserIdAndFactoryId/route.ts
@@ -1,15 +1,19 @@
-import { getReviewsByUserIdAndFactoryId } from '@/app/_features/review/api'
+import { getReviewsByUserIdAndFactoryId } from '@/app/_features/review/api/getReviewsByUserIdAndFactoryId'
 
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url)
   const userId = searchParams.get('userId')
   const factoryId = searchParams.get('factoryId')
-  // console.log({ userId, factoryId })
-  if (userId && factoryId) {
-    const res = await getReviewsByUserIdAndFactoryId({ userId, factoryId })
-    console.log(res)
-
-    return Response.json(res)
+  if (!userId || !factoryId) {
+    return Response.json({
+      status: 403,
+      message: 'userId and factoryId are required',
+    })
   }
-  return
+  try {
+    const res = await getReviewsByUserIdAndFactoryId({ userId, factoryId })
+    return Response.json(res)
+  } catch (error: any) {
+    throw new Error(error)
+  }
 }
